restful/hawkeye/requirements.txt
amqp==2.2.2
asgi-redis==1.4.2
asgiref==1.1.2
attrs==17.2.0
autobahn==17.9.3
Automat==0.6.0
billiard==3.5.0.3
celery==4.1.0
certifi==2017.7.27.1
channels==1.1.8
chardet==3.0.4
constantly==15.1.0
daphne==1.3.0
Django==2.0.1
django-filter==1.0.4
django-guardian==1.4.9
django-redis==4.8.0
djangorestframework==3.7.7
djangorestframework-bulk==0.2.1
djangorestframework-filters==0.10.1
djangorestframework-jwt==1.11.0
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
kombu==4.1.0
msgpack-python==0.4.8
olefile==0.44
Pillow==4.2.0
psycopg2==2.7.1
pycrypto==2.6.1
PyJWT==1.5.3
pytz==2017.2
redis==2.10.6
requests==2.18.4
rest-condition==1.0.3
six==1.11.0
Twisted==17.9.0
txaio==2.8.2
urllib3==1.22
vine==1.1.4
zope.interface==4.4.3
django-rest-swagger==2.1.2
django-celery-beat==1.1.0
pdfkit==0.6.1
sqlparse==0.2.4
django-debug-toolbar==1.9.1
gunicorn==19.7.1
gevent==1.2.2
paramiko==2.4.0